UPD separate service's request auth from services
- Creates separate serviceAuthenticator package providing
  implementation of runtime.ClientAuthInfoWriter interface
  for internal services using cert and key to authorize its requests to other
  internal services.
